Agrega Grids de cards cuadradas faltantes
